💄Move trigger key settings into Keybindings tab, delete dead code
