+++ Loop/Settings/Views/PreviewSettingsView.swift
@@ -43,36 +43,3 @@ struct PreviewSettingsView: View {
         .formStyle(.grouped)
     }
 }
-
-
-//                VStack(spacing: 10) {
-//
-//                    // Loop's preview window preview
-//                    ZStack {
-//                        RoundedRectangle(cornerRadius: 5)
-//                            .stroke(.secondary.opacity(0.35), lineWidth: 0.5)
-//                            .background(VisualEffectView(material: .sidebar, blendingMode: .behindWindow))
-//                            .clipShape(RoundedRectangle(cornerRadius: 5))
-//
-//                        ZStack {    // Grid Background
-//                            VStack {
-//                                Spacer()
-//                                ForEach(1...10, id: \.self) {_ in
-//                                    Divider()
-//                                    Spacer()
-//                                }
-//                            }
-//
-//                            HStack {
-//                                Spacer()
-//                                ForEach(1...27, id: \.self) {_ in
-//                                    Divider()
-//                                    Spacer()
-//                                }
-//                            }
-//                        }
-//
-//                        PreviewView(previewMode: true)
-//                            .padding(.horizontal, 30)
-//                    }
-//                    .frame(height: 150)
